+++ src/webviews/apps/commitDetails/actions.ts
@@ -562,6 +562,13 @@ export class CommitDetailsActions {
 		gitActions.discardUnstagedFiles(this.state.error, this.services.repository, repoPath);
 	}
 
+	discardStagedFiles(): void {
+		const repoPath = this.getRepoPath();
+		if (!repoPath) return;
+
+		gitActions.discardStagedFiles(this.state.error, this.services.repository, repoPath);
+	}
+
 	openConflictChanges(file: GitFileChangeShape, side: 'current' | 'incoming'): void {
 		void this.services.repository.openConflictChanges(file, side);
 	}