+++ src/webviews/rebase/rebaseWebviewProvider.ts
@@ -515,7 +515,7 @@ export class RebaseWebviewProvider implements Disposable {
 
 		if (toCheckoutOurs.length) {
 			try {
-				await svc.ops?.checkoutConflictedPaths?.(toCheckoutOurs, 'ours');
+				await svc.ops?.restore(toCheckoutOurs, { side: 'ours' });
 			} catch (ex) {
 				failures.push({ paths: toCheckoutOurs, reason: ex });
 				toCheckoutOurs.length = 0;
@@ -523,7 +523,7 @@ export class RebaseWebviewProvider implements Disposable {
 		}
 		if (toCheckoutTheirs.length) {
 			try {
-				await svc.ops?.checkoutConflictedPaths?.(toCheckoutTheirs, 'theirs');
+				await svc.ops?.restore(toCheckoutTheirs, { side: 'theirs' });
 			} catch (ex) {
 				failures.push({ paths: toCheckoutTheirs, reason: ex });
 				toCheckoutTheirs.length = 0;
@@ -591,10 +591,10 @@ export class RebaseWebviewProvider implements Disposable {
 				await svc.staging?.removeFile(path, { force: true });
 				return;
 			case 'take-ours':
-				await svc.ops?.checkoutConflictedPath?.(path, 'ours');
+				await svc.ops?.restore(path, { side: 'ours' });
 				break;
 			case 'take-theirs':
-				await svc.ops?.checkoutConflictedPath?.(path, 'theirs');
+				await svc.ops?.restore(path, { side: 'theirs' });
 				break;
 			case 'unsupported':
 				throw new Error(`Cannot take ${resolution} side for conflict status ${status}`);